Use ACL to filter access (and returing objects) in service methods.
filterSecurityInterceptor filters access to webservice methods by ROLE (just as before)
methods are filtered using methodSecurity and pre-post-annotations requiring permissions

pre-post-annotations and spring-security-acl AclVoters implementation requires aspectj.
However, this dependency is not declared in the bundle manifest until spring security 3.0.8 version, and this is required for spring security to work in OSGi.
Thus, we require spring security 3.0.8 version (which leads to a dependency upgrade explained afterwards)

Configuration for Spring aclService requires a database and an aclCache.
Default implementation of the acl database uses HSQLDB. DB schema is now created using an SQL script (aclSchema.sql)
Spring has also helpers to use ehcache for the aclCache service. ehcache is configured by ehcache-failsafe.xml, which should be improved for other environments different than demonstration. According to ehcache documentation ehcache.xml should be used as a strarting point, however it does not work with the version of ehcache we are using.

TODO: Solve following error in MethodSecurityInterceptor:
org.springframework.security.authentication.AuthenticationCredentialsNotFoundException: An Authentication object was not found in the SecurityContext
	at org.springframework.security.access.intercept.AbstractSecurityInterceptor.credentialsNotFound(AbstractSecurityInterceptor.java:325)
	at org.springframework.security.access.intercept.AbstractSecurityInterceptor.beforeInvocation(AbstractSecurityInterceptor.java:196)
	at org.springframework.security.access.intercept.aopalliance.MethodSecurityInterceptor.invoke(MethodSecurityInterceptor.java:64)
	...

Dependency upgrade notes:

* spring-security 3.0.5 --> 3.0.8
spring-security-acl AclVoters implementation requires aspectj but this dependency is not declared until version 3.0.8.
spring security 3.0.8 is able to use aspectj in OSGi, so we require it.
spring security 3.0.8 requires spring [3.0.7, 3.1) and aspectj [1.6.0, 1.7).

* spring, spring-web 3.0.5 --> 3.0.7
spring-security 3.0.7 requires spring 3.0.7
spring 3.0.7 includes spring-web 3.0.7
spring-web 3.0.7 requires http version 2.2.6 which in turn requires jetty 7.5.4.v20111024

spring 3.0.7 is buldleled as a karaf feature in karaf 2.2.6 features. We are currently using karaf 2.2.2-fuse-04-06.
Currently both features are co-existing.